backend/requirements.txt
Django==5.1
psycopg2-binary==2.9.9
djangorestframework==3.15.1
django-cors-headers==4.3.1
dj-database-url==2.1.0
python-pptx==0.6.22
